# Conflicts:
#	src/segment/segment_instance_api.cpp

+++ src/api/arbor.h
@@ -1,3 +1,23 @@
+/**
+ * @file arbor.h
+ * Project: Arbor
+ *
+ * Copyright (C) 2026 Jean-Romain Roussel (r-lidar) <info @ r-lidar.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #ifndef API_H
 #define API_H
 
@@ -101,7 +121,7 @@ namespace arbor
     void segment_instance(PointCloud& core, const PointCloud& seeds, const settings::ArborParameters& params);
     void segment_semantic(PointCloud& core, const PointCloud& dtm,   const settings::ArborParameters& params);
 
-    std::vector<double> dist2root(const PointCloud& core, const PointCloud& dtm, const settings::GraphParameters& params);
+    std::vector<float> dist2root(const PointCloud& core, const PointCloud& dtm, const settings::GraphParameters& params);
   }
 
   namespace seeds
